What: Terraform apply fails for sensitive values
Fix:
Currently, there are two workarounds which are added in this fix.
Strategy 1 : uses count to counter the for-each evauluation of
 values which are transformed from sensitive values.It seems like
terraform some how marks the transformed object as sensitive , even
though there are no sensitive values in the transformed object. This can
be verified using nonsensitive() func.

Strategy 2: uses jsonencode() func to ensure the transformed object is
converted into string and vice-versa. This is probably a cleaner
approach than using count. The resource used for this apporach is
knows module.host.spot_instances. Separate variables and separate
resources for spot_keys are used to keep it simple.

Note: Both these approaches follow the initial requirement of not moving
the logic from child module to root module.